requirements.txt
pillow
torch
torchvision
numpy